+++ crates/recent_projects/Cargo.toml
@@ -31,7 +31,7 @@ smol.workspace = true
 task.workspace = true
 terminal_view.workspace = true
 ui.workspace = true
-ui_text_field.workspace = true
+ui_input.workspace = true
 util.workspace = true
 workspace.workspace = true
