+++ crates/recent_projects/src/dev_servers.rs
@@ -31,7 +31,7 @@ use ui::{
     prelude::*, Indicator, List, ListHeader, ListItem, Modal, ModalFooter, ModalHeader,
     RadioWithLabel, Tooltip,
 };
-use ui_text_field::{FieldLabelLayout, TextField};
+use ui_input::{FieldLabelLayout, TextField};
 use util::ResultExt;
 use workspace::{notifications::DetachAndPromptErr, AppState, ModalView, Workspace, WORKSPACE_DB};
