+++ crates/ui_input/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ui_text_field"
+name = "ui_input"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -9,7 +9,7 @@ license = "GPL-3.0-or-later"
 workspace = true
 
 [lib]
-path = "src/ui_text_field.rs"
+path = "src/ui_input.rs"
 
 [dependencies]
 editor.workspace = true
